requirements.txt
amqp==2.1.4
appdirs==1.4.3
billiard==3.5.0.2
celery==4.0.2
Django==1.8.3
django-celery-beat==1.0.1
django-celery-results==1.0.1
httplib2==0.10.3
kombu==4.0.2
oauth==1.0.1
oauth2==1.5.211
oauth2client==1.4.12
oauthlib==2.0.0
packaging==16.8
pyasn1==0.2.3
pyasn1-modules==0.0.8
PyJWT==1.5.0
pyparsing==2.2.0
python-openid==2.2.5
python-social-auth==0.2.12
pytz==2017.2
redis==2.10.5
requests==2.11.1
requests-oauth2==0.2.0
requests-oauthlib==0.5.0
rsa==3.4.2
six==1.10.0
vine==1.1.3
